mcp/remote-server-py/requirements.txt
mcp>=1.9.0
uvicorn>=0.29
PyJWT>=2.9
cryptography>=43.0
httpx>=0.27
